Don't parse floats as ints

+++ src/discord.ts
@@ -156,10 +156,10 @@ async function sendHookWithEmbeds(env: Env, ctx: ExecutionContext, hook: string,
 	// Try to preimptively ratelimit if needed
 	try {
 		const rateLimitRemaining = discordResponse.headers.get('X-RateLimit-Remaining')
-		if (rateLimitRemaining && parseInt(rateLimitRemaining) < 1) { // Maybe it returns less than 1 but > 0?
+		if (rateLimitRemaining && parseFloat(rateLimitRemaining) < 1) { // Maybe it returns less than 1 but > 0?
 			const rateLimitResetAfter = discordResponse.headers.get('X-RateLimit-Reset-After')
 			if (rateLimitResetAfter) {
-				const resetAfter = parseInt(rateLimitResetAfter)
+				const resetAfter = parseFloat(rateLimitResetAfter)
 				if (resetAfter > 0) {
 					console.log(`Ratelimited! Sleeping for ${resetAfter} seconds...`)
 					logtail({
